# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: PlaceMIT - College Placement Community Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /students:
    get:
      operationId: listStudents
      tags: [students]
      summary: List all student profiles
      responses:
        "200":
          description: List of students
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Student"
    post:
      operationId: createStudent
      tags: [students]
      summary: Create a student profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudentInput"
      responses:
        "201":
          description: Created student
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
  /students/{id}:
    get:
      operationId: getStudent
      tags: [students]
      summary: Get a student profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Student details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
    patch:
      operationId: updateStudent
      tags: [students]
      summary: Update a student profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudentUpdate"
      responses:
        "200":
          description: Updated student
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"

  # Community Posts
  /posts:
    get:
      operationId: listPosts
      tags: [posts]
      summary: List community posts
      parameters:
        - name: category
          in: query
          schema:
            type: string
            enum: [general, interview-prep, resume-help, offers, advice]
      responses:
        "200":
          description: List of posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Post"
    post:
      operationId: createPost
      tags: [posts]
      summary: Create a community post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostInput"
      responses:
        "201":
          description: Created post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
  /posts/{id}:
    get:
      operationId: getPost
      tags: [posts]
      summary: Get a community post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Post details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
    delete:
      operationId: deletePost
      tags: [posts]
      summary: Delete a post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /posts/{id}/upvote:
    post:
      operationId: upvotePost
      tags: [posts]
      summary: Upvote a post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"

  # Comments
  /posts/{postId}/comments:
    get:
      operationId: listComments
      tags: [comments]
      summary: List comments on a post
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Comment"
    post:
      operationId: createComment
      tags: [comments]
      summary: Add a comment to a post
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommentInput"
      responses:
        "201":
          description: Created comment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"

  # Placement Experiences
...
components:
  schemas:
...
    Student:
      type: object
      required: [id, name, rollNo, branch, year, cgpa, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        rollNo:
          type: string
        branch:
          type: string
        year:
          type: integer
        cgpa:
          type: number
        bio:
          type: ["string", "null"]
        skills:
          type: ["string", "null"]
        linkedinUrl:
          type: ["string", "null"]
        githubUrl:
          type: ["string", "null"]
        resumeUrl:
          type: ["string", "null"]
        placedAt:
          type: ["string", "null"]
        packageOffered:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        createdAt:
          type: string
    StudentInput:
      type: object
      required: [name, rollNo, branch, year, cgpa]
      properties:
        name:
          type: string
        rollNo:
          type: string
        branch:
          type: string
        year:
          type: integer
        cgpa:
          type: number
        bio:
          type: string
        skills:
          type: string
        linkedinUrl:
          type: string
        githubUrl:
          type: string
        resumeUrl:
          type: string
        packageOffered:
          type: string
        avatarUrl:
          type: string
    StudentUpdate:
      type: object
      properties:
        bio:
          type: string
        skills:
          type: string
        linkedinUrl:
          type: string
        githubUrl:
          type: string
        resumeUrl:
          type: string
        placedAt:
          type: string
        packageOffered:
          type: string

    # Post schemas
    Post:
      type: object
      required: [id, title, content, authorName, category, upvotes, commentsCount, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: string
        authorName:
          type: string
        authorAvatarUrl:
          type: ["string", "null"]
        category:
          type: string
          enum: [general, interview-prep, resume-help, offers, advice]
        upvotes:
          type: integer
        commentsCount:
          type: integer
        createdAt:
          type: string
    PostInput:
      type: object
      required: [title, content, authorName, category]
      properties:
        title:
          type: string
        content:
          type: string
        authorName:
          type: string
        authorAvatarUrl:
          type: string
        category:
          type: string
          enum: [general, interview-prep, resume-help, offers, advice]

    # Comment schemas
    Comment:
      type: object
      required: [id, postId, content, authorName, createdAt]
      properties:
        id:
          type: integer
        postId:
          type: integer
        content:
          type: string
        authorName:
          type: string
        authorAvatarUrl:
          type: ["string", "null"]
        createdAt:
          type: string
    CommentInput:
      type: object
      required: [content, authorName]
      properties:
        content:
          type: string
        authorName:
          type: string
        authorAvatarUrl:
          type: string

    # Experience schemas
...