# lib/api-zod/src/generated/types/student.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PlaceMIT - College Placement Community Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Student {
  id: number;
  name: string;
  rollNo: string;
  branch: string;
  year: number;
  cgpa: number;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  skills?: string | null;
  /** @nullable */
  linkedinUrl?: string | null;
  /** @nullable */
  githubUrl?: string | null;
  /** @nullable */
  resumeUrl?: string | null;
  /** @nullable */
  placedAt?: string | null;
  /** @nullable */
  packageOffered?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  createdAt: string;
}