# lib/api-zod/src/generated/types/studentInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PlaceMIT - College Placement Community Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface StudentInput {
  name: string;
  rollNo: string;
  branch: string;
  year: number;
  cgpa: number;
  bio?: string;
  skills?: string;
  linkedinUrl?: string;
  githubUrl?: string;
  resumeUrl?: string;
  packageOffered?: string;
  avatarUrl?: string;
}